fix: prepay API 기능 개발
