+++ src/main/java/com/jangburich/presentation/prepay/dto/request/PrepayRequest.java
@@ -3,7 +3,6 @@ package com.jangburich.presentation.prepay.dto.request;
 public record PrepayRequest(
         Long storeId,
         Long teamId,
-        int prepayAmount,
-        int personalAllocatedAmount
+        int prepayAmount
 ) {
 }
